added docs and license headers
