+++ ethcore/src/blockchain/best_block.rs
@@ -1,23 +1,30 @@
+// Copyright 2015, 2016 Ethcore (UK) Ltd.
+// This file is part of Parity.
+
+// Parity is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Parity is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
+
 use util::hash::H256;
 use util::uint::U256;
 use header::BlockNumber;
 
-/// Information about best block gathered together
+/// Best block info.
 #[derive(Default)]
 pub struct BestBlock {
+	/// Best block hash.
 	pub hash: H256,
+	/// Best block number.
 	pub number: BlockNumber,
+	/// Best block total difficulty.
 	pub total_difficulty: U256
 }
-
-impl BestBlock {
-	pub fn new() -> BestBlock { Default::default() }
-}
-		
-		//BestBlock {
-			//hash: H256::new(),
-			//number: 0,
-			//total_difficulty: U256::from(0)
-		//}
-	//}
-//}
